+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 * Added: System information on new remoting connection
+* Fixed: Fix the Update-Workspace command to write a config file without the UTF-8 BOM header
+* Fixed: Fix the Update-Workspace command if no config file is present
 
 ## 4.4.0 - 2022-12-27